+++ entry/entry-modular.js
@@ -1,0 +1,11 @@
+import Sortable from '../src/Sortable.js';
+import Swap from '../plugins/Swap.js';
+import MultiDrag from '../plugins/MultiDrag.js';
+
+export default Sortable;
+
+export {
+	Sortable,
+	Swap,
+	MultiDrag
+};
